requirements.txt
Django
djangorestframework
requests
pydantic
djangorestframework-simplejwt
hvac